server/requirement.txt
fastapi == 0.115.12
uvicorn == 0.34.2
motor == 3.7.0
python-dotenv == 1.1.0
openai
pydantic == 2.11.3
google-generativeai 
sentry-sdk[fastapi]